requirements.txt
python-telegram-bot==21.0
librosa==0.10.1
numpy==1.26.4
scipy==1.11.4
matplotlib==3.8.2
pandas==2.1.4
python-dotenv==1.0.0